Traffic light code
A code programmed to create a pedestrian priority traffic light system. This code modifies the idea of regular pedestrian crossing by giving priority to pedestrians instead of vehicles. The code contains a loop of regular traffic lights which is interrupted if a pedestrian is detected. Once detected, the code displays a yellow signal to the vehicles and then red, signalling a full stop. It also makes a buzzer sound once the stop has occurred making it friendly and easy for people with visual disabilities.